market_mcp/requirements.txt
mcp>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
google-cloud-bigquery>=3.11.0
httpx>=0.25.0
pydantic>=2.0.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi

opentelemetry-exporter-gcp-trace>=1.6.0
google-cloud-trace>=1.11.0
python-jose[cryptography]>=3.3.0
prometheus-fastapi-instrumentator>=6.1.0
redis>=5.0.0


opentelemetry-instrumentation-httpx

opentelemetry-instrumentation-redis
